// ui/src/app/diagrams/models/resizer-offsets.model.ts
export interface ResizerOffset {
  x: number;
  y: number;
}

export interface ResizerOffsets {
  top: ResizerOffset;
  left: ResizerOffset;
  bottom: ResizerOffset;
  right: ResizerOffset;
  topLeft: ResizerOffset;
  topRight: ResizerOffset;
  bottomLeft: ResizerOffset;
  bottomRight: ResizerOffset;
}